Aggiunta log e timer al Trasportatore + Fix vari
Aggiunto sistema di log e timer al Trasportatore + Vari fix + Rimossa funzione prodottiOrdine

+++ servers/Trasportatore/metodi/rimuoviCorriere.cpp
@@ -10,6 +10,11 @@ bool rimuoviCorriere(int trasporterID, int courierID)
     
     try
     {
+        sprintf(comando, "SELECT * FROM corriere WHERE id = %d AND azienda = %d", courierID, trasporterID);
+        res = db.ExecSQLtuples(comando);
+        if (PQntuples(res) == 0) return false;  // Se il corriere non è nel sistema, impedisce l'operazione
+        
+        PQclear(res);
         sprintf(comando, "DELETE FROM corriere WHERE id = %d AND azienda = %d", courierID, trasporterID);
         res = db.ExecSQLcmd(comando);
         PQclear(res);
